Try refactor project structure
