+++ .gitignore
@@ -23,3 +23,4 @@ vendor/
 # IDEs directories
 .idea
 .vscode
+.run/
