+++ internal/agent/entities/snapshot.go
@@ -1,0 +1,6 @@
+package entities
+
+type Snapshot struct {
+	Gauges   map[string]float64
+	Counters map[string]int64
+}
