+++ internal/infra/config/agent/agent.go
@@ -1,4 +1,4 @@
-package config
+package agent
 
 import (
 	"flag"
