+++ internal/infra/config/agent/agent_test.go
@@ -1,4 +1,4 @@
-package config
+package agent
 
 import (
 	"flag"
