+++ internal/infra/config/server/server.go
@@ -1,4 +1,4 @@
-package config
+package server
 
 import (
 	"flag"
