+++ internal/infra/config/server/server_test.go
@@ -1,4 +1,4 @@
-package config
+package server
 
 import (
 	"flag"
