+++ internal/infra/logging/adapters.go
@@ -1,4 +1,4 @@
-package adapters
+package logging
 
 import (
 	"os"
