+++ internal/infra/logging/interfaces.go
@@ -1,6 +1,5 @@
-package interfaces
+package logging
 
-// todo: read about clean architecture in Go for web projects
 type ILogger interface {
 	Info(args ...interface{})
 	Infof(format string, args ...interface{})
