+++ internal/server/usecases/core.go
@@ -1,4 +1,4 @@
-package handlers
+package usecases
 
 import (
 	"errors"
@@ -6,11 +6,11 @@ import (
 	"net/http"
 	"path/filepath"
 
-	"github.com/matthiasBT/monitoring/internal/storage"
+	"github.com/matthiasBT/monitoring/internal/server/entities"
 )
 
 func UpdateMetric(w http.ResponseWriter, c *BaseController, params map[string]string) {
-	metricUpdate := storage.MetricUpdate{
+	metricUpdate := entities.MetricUpdate{
 		Type:  params["type"],
 		Name:  params["name"],
 		Value: params["value"],
@@ -18,11 +18,11 @@ func UpdateMetric(w http.ResponseWriter, c *BaseController, params map[string]st
 	err := metricUpdate.Validate()
 	if err != nil {
 		switch {
-		case errors.Is(err, storage.ErrInvalidMetricType):
+		case errors.Is(err, entities.ErrInvalidMetricType):
 			w.WriteHeader(http.StatusBadRequest)
-		case errors.Is(err, storage.ErrMissingMetricName):
+		case errors.Is(err, entities.ErrMissingMetricName):
 			w.WriteHeader(http.StatusNotFound)
-		case errors.Is(err, storage.ErrInvalidMetricVal):
+		case errors.Is(err, entities.ErrInvalidMetricVal):
 			w.WriteHeader(http.StatusBadRequest)
 		default:
 			w.WriteHeader(http.StatusInternalServerError)
@@ -30,16 +30,16 @@ func UpdateMetric(w http.ResponseWriter, c *BaseController, params map[string]st
 		w.Write([]byte(err.Error()))
 		return
 	} else {
-		c.stor.Add(metricUpdate)
+		c.Stor.Add(metricUpdate)
 		w.WriteHeader(http.StatusOK)
 		return
 	}
 }
 
 func GetMetric(w http.ResponseWriter, c *BaseController, params map[string]string) {
-	val, err := c.stor.Get(params["type"], params["name"])
+	val, err := c.Stor.Get(params["type"], params["name"])
 	if err != nil {
-		if errors.Is(err, storage.ErrUnknownMetricName) || errors.Is(err, storage.ErrInvalidMetricType) {
+		if errors.Is(err, entities.ErrUnknownMetricName) || errors.Is(err, entities.ErrInvalidMetricType) {
 			w.WriteHeader(http.StatusNotFound)
 		} else {
 			w.WriteHeader(http.StatusInternalServerError)
@@ -54,11 +54,11 @@ func GetMetric(w http.ResponseWriter, c *BaseController, params map[string]strin
 }
 
 func GetAllMetrics(w http.ResponseWriter, c *BaseController, templateName string) {
-	data := c.stor.GetAll()
-	path := filepath.Join(c.templatePath, templateName)
+	data := c.Stor.GetAll()
+	path := filepath.Join(c.TemplatePath, templateName)
 	tmpl := template.Must(template.ParseFiles(path))
 	err := tmpl.Execute(w, data)
 	if err != nil {
-		c.logger.Fatal(err)
+		c.Logger.Fatal(err)
 	}
 }
